+++ temp_cata/BetterBlizzPlates.lua
@@ -5184,8 +5184,18 @@ local function HandleNamePlateAdded(unit)
         if BetterBlizzPlatesDB.changeNameplateBorderSize then
             if info.isTarget then
                 frame.healthBar:SetBorderSize(BetterBlizzPlatesDB.nameplateTargetBorderSize)
+                if BetterBlizzPlatesDB.castBarPixelBorder then
+                    if frame.CastBar.SetBorderSize then
+                        frame.CastBar:SetBorderSize(BetterBlizzPlatesDB.nameplateTargetBorderSize)
+                    end
+                end
             else
                 frame.healthBar:SetBorderSize(BetterBlizzPlatesDB.nameplateBorderSize)
+                if BetterBlizzPlatesDB.castBarPixelBorder then
+                    if frame.CastBar.SetBorderSize then
+                        frame.CastBar:SetBorderSize(BetterBlizzPlatesDB.nameplateBorderSize)
+                    end
+                end
             end
         end
     end
@@ -5592,8 +5602,18 @@ function BBP.RefreshAllNameplates()
         if BetterBlizzPlatesDB.changeNameplateBorderSize then
             if info.isTarget then
                 frame.healthBar:SetBorderSize(BetterBlizzPlatesDB.nameplateTargetBorderSize)
+                if BetterBlizzPlatesDB.castBarPixelBorder then
+                    if frame.CastBar.SetBorderSize then
+                        frame.CastBar:SetBorderSize(BetterBlizzPlatesDB.nameplateTargetBorderSize)
+                    end
+                end
             else
                 frame.healthBar:SetBorderSize(BetterBlizzPlatesDB.nameplateBorderSize)
+                if BetterBlizzPlatesDB.castBarPixelBorder then
+                    if frame.CastBar.SetBorderSize then
+                        frame.CastBar:SetBorderSize(BetterBlizzPlatesDB.nameplateBorderSize)
+                    end
+                end
             end
         end