+++ changelog.md
@@ -1,5 +1,7 @@
 # BetterBlizzPlates 1.8.4j
 ## Retail
+### New
+- Added Mes's Profile. Thank you <3
 ### Tweak
 - Add unique pet icons for Class Indicator for Hunter pet families. Huge thanks to Stroold @ Discord for putting this list together <3
 - Blitz preset profile adjusted to show pvp buffs by default too.
@@ -7,6 +9,8 @@
 ## All versions
 ### New
 - New setting for nameplate background color which makes it solid and allows you to properly put any color as background which was limited before.
+### Tweak
+- Change nameplate border size now also resizes border on castbar and castbar icon if enabled and not just healthbar.
 
 # BetterBlizzPlates 1.8.4i
 ## Retail